Created OpenF1APIService to ingest data from OpenF1 API

+++ src/main/java/com/f1telemetry/race_telemetry_analyzer/controller/DriverController.java
@@ -2,8 +2,12 @@ package com.f1telemetry.race_telemetry_analyzer.controller;
 
 import com.f1telemetry.race_telemetry_analyzer.model.Driver;
 import com.f1telemetry.race_telemetry_analyzer.service.DriverService;
-import org.springframework.beans.factory.annotation.Autowired;
+import com.f1telemetry.race_telemetry_analyzer.service.OpenF1ApiService;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -13,8 +17,13 @@ import java.util.Optional;
 @RequestMapping("/api/drivers")
 public class DriverController {
 
-    @Autowired
-    private DriverService driverService;
+
+    private final DriverService driverService;
+    private static final Logger logger = LoggerFactory.getLogger(OpenF1ApiController.class);
+
+    public DriverController(DriverService driverService, OpenF1ApiService openF1ApiService) {
+        this.driverService = driverService;
+    }
 
     @GetMapping
     public List<Driver> getAllDrivers() {
@@ -28,12 +37,21 @@ public class DriverController {
     }
 
     @PostMapping
-    public Driver addDriver(@RequestBody Driver driver) {
-        return driverService.addDriver(driver);
+    public ResponseEntity<Driver> addDriver(@Validated @RequestBody Driver driver) {
+        logger.info("Adding a new driver: {}", driver.getName());
+        Driver savedDriver = driverService.addDriver(driver);
+        return ResponseEntity.status(HttpStatus.CREATED).body(savedDriver);
     }
 
     @DeleteMapping("/{id}")
     public ResponseEntity<Void> deleteDriver(@PathVariable String id) {
+        logger.info("Deleting driver with id: {}", id);
+        Optional<Driver> driver = driverService.getDriverById(id);
+        if (driver.isEmpty()) {
+            logger.warn("Driver with id {} not found", id);
+            return ResponseEntity.notFound().build();
+        }
+
         driverService.deleteDriver(id);
         return ResponseEntity.noContent().build();
     }
